+++ CHANGELOG.md
@@ -5,7 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-*No unreleased changes*
+### Changed
+- Improve the prevention of a pop-up window outside the screen
 
 ## [1.0.0] - 2018-07-01
 ### Added